feat: require teacher auth when operate class info

+++ src/app/api/class/delete/route.ts
@@ -6,8 +6,13 @@ const prisma = new PrismaClient();
 export async function POST(request: NextRequest) {
   const info = await request.json();
   try {
-    const user = await prisma.class.delete({ where: { name: info.name } });
-    return NextResponse.json({ message: "ok"});
+    const teacher = await prisma.teacher.findUniqueOrThrow({ where: { name: info.teacher_name } });
+    if(teacher.pass == info.teacher_pass){
+      const user = await prisma.class.delete({ where: { name: info.name } });
+      return NextResponse.json({ message: "ok"});
+    }else{
+      return NextResponse.json({ message: "error" }, { status: 500 });
+    }
   } catch (e) {
     return NextResponse.json({ message: "error" }, { status: 500 });
   }
